feat : refacto des conditions sur les unlink de MenuController et PlatController problème sur upload , vue menu détaillé terminé , import des images menus

+++ app/controllers/MenuController.php
@@ -341,7 +341,10 @@ class MenuController{
                 header('Location: /menus/edit/'. $id .'?error=' . urlencode($error));
                 exit();
             }
-            unlink(__DIR__ . "/../../public/" . $menu['chemin']);
+
+            if($menu['chemin'] && file_exists(__DIR__ . "/../../public/" . $menu['chemin'])){
+                unlink(__DIR__ . "/../../public/" . $menu['chemin']);
+            }
             $this->menus->deleteImage($id);
             move_uploaded_file($source, $destination);
